Refactor project packages architecture
