+++ TagsCloud/TagsCloudVisualization/CloudLayouter/VectorsGenerator/IVectorsGenerator.cs
@@ -1,0 +1,9 @@
+﻿using System.Drawing;
+
+namespace TagsCloudVisualization.CloudLayouter.VectorsGenerator
+{
+    public interface IVectorsGenerator
+    {
+        Point GetNextVector();
+    }
+}
