+++ TagsCloud/TagsCloudVisualization/Drawable/Tags/Factory/ITagDrawableFactory.cs
@@ -1,0 +1,7 @@
+﻿namespace TagsCloudVisualization.Drawable.Tags.Factory
+{
+    public interface ITagDrawableFactory
+    {
+        TagDrawable Create(Tag tag);
+    }
+}
