+++ TagsCloud/TagsCloudVisualization/Drawable/Tags/Settings/ITagDrawableSettingsProvider.cs
@@ -1,0 +1,10 @@
+﻿using TagsCloudVisualization.Drawable.Tags.Settings.TagColorGenerator;
+
+namespace TagsCloudVisualization.Drawable.Tags.Settings
+{
+    public interface ITagDrawableSettingsProvider
+    {
+        FontSettings Font { get; }
+        ITagColorGenerator ColorGenerator { get; }
+    }
+}
