+++ TagsCloud/TagsCloudVisualization/Drawable/Tags/Settings/TagColorGenerator/ITagColorGenerator.cs
@@ -1,0 +1,9 @@
+﻿using System.Drawing;
+
+namespace TagsCloudVisualization.Drawable.Tags.Settings.TagColorGenerator
+{
+    public interface ITagColorGenerator
+    {
+        Color Generate(Tag tag);
+    }
+}
