+++ TagsCloud/TagsCloudVisualization/Drawable/Tags/Settings/TagDrawableSettingsProvider.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using TagsCloudDrawer.ColorGenerators;
+using TagsCloudVisualization.Drawable.Tags.Settings.TagColorGenerator;
+
+namespace TagsCloudVisualization.Drawable.Tags.Settings
+{
+    public class TagDrawableSettingsProvider : ITagDrawableSettingsProvider
+    {
+        public FontSettings Font { get; init; } = new();
+
+        public ITagColorGenerator ColorGenerator { get; init; } =
+            new RandomTagColorGenerator(new RandomColorGenerator(new Random()));
+    }
+}
