--- TagsCloud/TagsCloudVisualization/DrawableFactory/ITagDrawableFactory.cs
@@ -1,9 +1,0 @@
-﻿using TagsCloudVisualization.Drawable;
-
-namespace TagsCloudVisualization.DrawableFactory
-{
-    public interface ITagDrawableFactory
-    {
-        TagDrawable Create(Tag tag);
-    }
-}
