--- TagsCloud/TagsCloudVisualization/DrawerSettingsProvider/ITagDrawableSettingsProvider.cs
@@ -1,10 +1,0 @@
-﻿using TagsCloudVisualization.DrawerSettingsProvider.TagColorGenerator;
-
-namespace TagsCloudVisualization.DrawerSettingsProvider
-{
-    public interface ITagDrawableSettingsProvider
-    {
-        FontSettings Font { get; }
-        ITagColorGenerator ColorGenerator { get; }
-    }
-}
