--- TagsCloud/TagsCloudVisualization/DrawerSettingsProvider/TagColorGenerator/ITagColorGenerator.cs
@@ -1,9 +1,0 @@
-﻿using System.Drawing;
-
-namespace TagsCloudVisualization.DrawerSettingsProvider.TagColorGenerator
-{
-    public interface ITagColorGenerator
-    {
-        Color Generate(Tag tag);
-    }
-}
