--- TagsCloud/TagsCloudVisualization/DrawerSettingsProvider/TagColorGenerator/RandomTagColorGenerator.cs
@@ -1,17 +1,0 @@
-﻿using System.Drawing;
-using TagsCloudDrawer.ColorGenerators;
-
-namespace TagsCloudVisualization.DrawerSettingsProvider.TagColorGenerator
-{
-    public class RandomTagColorGenerator : ITagColorGenerator
-    {
-        private readonly IColorGenerator _generator;
-
-        public RandomTagColorGenerator(IColorGenerator generator)
-        {
-            _generator = generator;
-        }
-
-        public Color Generate(Tag tag) => _generator.Generate();
-    }
-}
