+++ TagsCloud/TagsCloudVisualization/Extensions/RectangleExtensions.cs
@@ -2,7 +2,7 @@
 
 namespace TagsCloudVisualization.Extensions
 {
-    public static class RectangleExtensions
+    internal static class RectangleExtensions
     {
         public static Point GetCenter(this Rectangle rect)
             => new(rect.X + rect.Width / 2, rect.Y + rect.Height / 2);
