+++ TagsCloud/TagsCloudVisualization/Module/TagsCloudDrawerModule.cs
@@ -6,8 +6,8 @@ using TagsCloudDrawer.ImageCreator;
 using TagsCloudDrawer.ImageSaveService;
 using TagsCloudDrawer.ImageSettings;
 using TagsCloudVisualization.CloudLayouter;
-using TagsCloudVisualization.DrawableFactory;
-using TagsCloudVisualization.DrawerSettingsProvider;
+using TagsCloudVisualization.Drawable.Tags.Factory;
+using TagsCloudVisualization.Drawable.Tags.Settings;
 using TagsCloudVisualization.WordsPreprocessor;
 using TagsCloudVisualization.WordsProvider;
 using TagsCloudVisualization.WordsToTagsTransformers;
