+++ TagsCloud/TagsCloudVisualization/Module/TagsCloudDrawerModuleSettings.cs
@@ -5,8 +5,8 @@ using System.Linq;
 using TagsCloudDrawer.ImageSaveService;
 using TagsCloudDrawer.ImageSettings;
 using TagsCloudVisualization.CloudLayouter;
-using TagsCloudVisualization.DrawerSettingsProvider;
-using TagsCloudVisualization.DrawerSettingsProvider.TagColorGenerator;
+using TagsCloudVisualization.Drawable.Tags.Settings;
+using TagsCloudVisualization.Drawable.Tags.Settings.TagColorGenerator;
 using TagsCloudVisualization.WordsPreprocessor;
 
 namespace TagsCloudVisualization.Module
