--- TagsCloud/TagsCloudVisualization/VectorsGenerator/IVectorsGenerator.cs
@@ -1,9 +1,0 @@
-﻿using System.Drawing;
-
-namespace TagsCloudVisualization.VectorsGenerator
-{
-    public interface IVectorsGenerator
-    {
-        Point GetNextVector();
-    }
-}
